Split basic_csv_parser
